combine install and fetch cmd files
